Fix clippy warnings for Rust 1.97 compatibility
Fix 11 clippy errors found by CI (Rust 1.97):
- empty_line_after_doc_comments: orphaned doc comment → regular comment
- unnecessary_mut_passed: Parser::parse_sql takes &Dialect, not &mut Dialect
- unnecessary_map_or: .map_or(true, |x| pred(x)) → .is_none_or(|x| pred(x))
- unnecessary_cast: z as i64 → z (where z is already i64)
- redundant_closure: |n| object_name_str(n) → object_name_str
- enumerate_without_index: for (_i, x) → for x

+++ extension/src/engine/functions/builtin.rs
@@ -78,8 +78,8 @@ pub(crate) fn curdate_value() -> DbValue {
         .unwrap_or_default()
         .as_secs() as i64;
     let z = secs / 86400 + 719468;
-    let era = z as i64 / 146097;
-    let doe = z as i64 - era * 146097;
+    let era = z / 146097;
+    let doe = z - era * 146097;
     let yoe = (doe - doe / 1460 + doe / 36524 - doe / 146096) / 365;
     let y = yoe + era * 400;
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
